Allow generating Goal Comparisons
This separates out the logic from the Balanced PB comparison. The
Balanced PB algorithm balances the splits based on the history such that
the final time matches the Personal Best. However the algorithm is
flexible enough to be used to generate splits for other times as well,
as long as they are within the sum of the best segments and the sum of
worst segments. By separating out the goal generating algorithm, we
provide new APIs for generating custom comparisons for arbitrary goal
times. Furthermore this is now also exposed in the Run Editor for the
runner to use. In the Run Editor you only provide the goal time for the
currently selected timing method. If you want to provide both, you use
the feature twice.
